requirements.txt
wheel
gunicorn
flask
pyyaml
python-box
dash
pandas==2.2.2
sqlalchemy==2.0.32
dash-bootstrap-components
dash-svg
plotly
python-dotenv
matplotlib